Mac Launcher: Add some casts from CFStrings to NSStrings because technically the docs say to do that for toll-free bridged classes; add a missed argument type

+++ prboom2/src/MAC/ResolutionDataSource.h
@@ -29,7 +29,7 @@
 
 + (NSArray *)resolutions;
 
-- (id)comboBox:(NSComboBox *)box objectValueForItemAtIndex:i;
+- (id)comboBox:(NSComboBox *)box objectValueForItemAtIndex:(int)i;
 - (int)comboBox:(NSComboBox *)box indexOfItemWithStringValue:(NSString *)string;
 - (int)numberOfItemsInComboBox:(NSComboBox *)box;
